+++ effectful/CHANGELOG.md
@@ -1,5 +1,6 @@
 # effectful-2.6.1.0 (????-??-??)
 * Re-export `writeTMVar` from `stm-2.5.1.0` in `Effectful.Concurrent.STM`.
+* Add `cancelMany` to `Effectful.Concurrent.Async`.
 
 # effectful-core-2.6.0.0 (2025-06-13)
 * Adjust `generalBracket` with `base >= 4.21` to make use of the new exception